pyproject.toml
[project]
name = "youtube-mcp-server"
version = "0.1.0"
description = "MCP server for YouTube API integration with video search functionality"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "mcp>=1.6.0",
    "google-api-python-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.0.0",
    "youtube-transcript-api>=0.6.0",
]

[project.scripts]
youtube-mcp-server = "youtube_mcp_server.__main__:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["youtube_mcp_server"]

force-include = { "youtube_mcp_server/tools.json" = "youtube_mcp_server/tools.json" }

[tool.hatch.build.targets.sdist]
include = [
  "youtube_mcp_server/**",
  "README.md",
  "LICENSE",
]
